+++ wasm/src/utils.rs
@@ -4,7 +4,7 @@ use wasm_bindgen::prelude::wasm_bindgen;
 
 /// Compare point in Z-order
 /// => https://en.wikipedia.org/wiki/Z-order_curve
-pub fn cmp_zorder(lhs: &Point2<i64>, rhs: &Point2<i64>) -> cmp::Ordering {
+pub fn cmp_zorder(lhs: &Point2<i32>, rhs: &Point2<i32>) -> cmp::Ordering {
     let mut msd = 0;
 
     for dim in 0..2 {
@@ -19,7 +19,7 @@ pub fn cmp_zorder(lhs: &Point2<i64>, rhs: &Point2<i64>) -> cmp::Ordering {
     lhs[msd].cmp(&rhs[msd])
 }
 
-pub fn cmp_xy_order(lhs: &Point2<i64>, rhs: &Point2<i64>) -> cmp::Ordering {
+pub fn cmp_xy_order(lhs: &Point2<i32>, rhs: &Point2<i32>) -> cmp::Ordering {
     lhs.iter().cmp(rhs.iter())
 }
